Refactored 'CAST2FN' to 'CAST/Fortran'
