+++ skema/program_analysis/CAST/Fortran/ann_cast/ann_cast_pass_prototype.py.unused
@@ -1,7 +1,7 @@
 from functools import singledispatchmethod
 import typing
 
-from skema.program_analysis.CAST2FN.ann_cast.annotated_cast import *
+from skema.program_analysis.CAST.Fortran.ann_cast.annotated_cast import *
 
 
 class AnnCastPass:
