+++ skema/program_analysis/CAST/Fortran/ann_cast/variable_version_pass.py
@@ -3,7 +3,7 @@ from functools import singledispatchmethod
 
 from skema.model_assembly.metadata import VariableCreationReason, LambdaType
 from skema.model_assembly.networks import load_lambda_function
-from skema.program_analysis.CAST2FN.ann_cast.ann_cast_helpers import (
+from skema.program_analysis.CAST.Fortran.ann_cast.ann_cast_helpers import (
     ELSEBODY,
     IFBODY,
     IFEXPR,
@@ -32,8 +32,8 @@ from skema.program_analysis.CAST2FN.ann_cast.ann_cast_helpers import (
     is_func_def_main,
     specialized_global_name,
 )
-from skema.program_analysis.CAST2FN.ann_cast.annotated_cast import *
-from skema.program_analysis.CAST2FN.model.cast import (
+from skema.program_analysis.CAST.Fortran.ann_cast.annotated_cast import *
+from skema.program_analysis.CAST.Fortran.model.cast import (
     ScalarType,
     StructureType,
     ValueConstructor,
