+++ skema/program_analysis/CAST/Fortran/model/cast/__init__.py
@@ -14,26 +14,26 @@
 from __future__ import absolute_import
 
 # import models into model package
-from skema.program_analysis.CAST2FN.model.cast.assignment import Assignment
-from skema.program_analysis.CAST2FN.model.cast.ast_node import AstNode
-from skema.program_analysis.CAST2FN.model.cast.attribute import Attribute
-from skema.program_analysis.CAST2FN.model.cast.call import Call
-from skema.program_analysis.CAST2FN.model.cast.function_def import FunctionDef
-from skema.program_analysis.CAST2FN.model.cast.literal_value import LiteralValue
-from skema.program_analysis.CAST2FN.model.cast.loop import Loop
-from skema.program_analysis.CAST2FN.model.cast.model_break import ModelBreak
-from skema.program_analysis.CAST2FN.model.cast.model_continue import ModelContinue
-from skema.program_analysis.CAST2FN.model.cast.model_if import ModelIf
-from skema.program_analysis.CAST2FN.model.cast.model_import import ModelImport
-from skema.program_analysis.CAST2FN.model.cast.model_return import ModelReturn
-from skema.program_analysis.CAST2FN.model.cast.module import Module
-from skema.program_analysis.CAST2FN.model.cast.name import Name
-from skema.program_analysis.CAST2FN.model.cast.operator import Operator
-from skema.program_analysis.CAST2FN.model.cast.record_def import RecordDef
-from skema.program_analysis.CAST2FN.model.cast.scalar_type import ScalarType
-from skema.program_analysis.CAST2FN.model.cast.source_code_data_type import SourceCodeDataType
-from skema.program_analysis.CAST2FN.model.cast.source_ref import SourceRef
-from skema.program_analysis.CAST2FN.model.cast.structure_type import StructureType
-from skema.program_analysis.CAST2FN.model.cast.value_constructor import ValueConstructor
-from skema.program_analysis.CAST2FN.model.cast.var import Var
-from skema.program_analysis.CAST2FN.model.cast.var_type import VarType
+from skema.program_analysis.CAST.Fortran.model.cast.assignment import Assignment
+from skema.program_analysis.CAST.Fortran.model.cast.ast_node import AstNode
+from skema.program_analysis.CAST.Fortran.model.cast.attribute import Attribute
+from skema.program_analysis.CAST.Fortran.model.cast.call import Call
+from skema.program_analysis.CAST.Fortran.model.cast.function_def import FunctionDef
+from skema.program_analysis.CAST.Fortran.model.cast.literal_value import LiteralValue
+from skema.program_analysis.CAST.Fortran.model.cast.loop import Loop
+from skema.program_analysis.CAST.Fortran.model.cast.model_break import ModelBreak
+from skema.program_analysis.CAST.Fortran.model.cast.model_continue import ModelContinue
+from skema.program_analysis.CAST.Fortran.model.cast.model_if import ModelIf
+from skema.program_analysis.CAST.Fortran.model.cast.model_import import ModelImport
+from skema.program_analysis.CAST.Fortran.model.cast.model_return import ModelReturn
+from skema.program_analysis.CAST.Fortran.model.cast.module import Module
+from skema.program_analysis.CAST.Fortran.model.cast.name import Name
+from skema.program_analysis.CAST.Fortran.model.cast.operator import Operator
+from skema.program_analysis.CAST.Fortran.model.cast.record_def import RecordDef
+from skema.program_analysis.CAST.Fortran.model.cast.scalar_type import ScalarType
+from skema.program_analysis.CAST.Fortran.model.cast.source_code_data_type import SourceCodeDataType
+from skema.program_analysis.CAST.Fortran.model.cast.source_ref import SourceRef
+from skema.program_analysis.CAST.Fortran.model.cast.structure_type import StructureType
+from skema.program_analysis.CAST.Fortran.model.cast.value_constructor import ValueConstructor
+from skema.program_analysis.CAST.Fortran.model.cast.var import Var
+from skema.program_analysis.CAST.Fortran.model.cast.var_type import VarType
