+++ skema/program_analysis/CAST/Fortran/model/cast/model_return.py
@@ -14,7 +14,7 @@ import pprint
 import re  # noqa: F401
 
 import six
-from skema.program_analysis.CAST2FN.model.cast.ast_node import AstNode  # noqa: F401,E501
+from skema.program_analysis.CAST.Fortran.model.cast.ast_node import AstNode  # noqa: F401,E501
 
 class ModelReturn(AstNode):
     """NOTE: This class is auto generated by the swagger code generator program.
