+++ skema/program_analysis/CAST/Fortran/visitors/cast_function_call_visitor.py
@@ -1,9 +1,9 @@
 import typing
 from functools import singledispatchmethod
 
-from skema.program_analysis.CAST2FN.visitors.cast_visitor import CASTVisitor
+from skema.program_analysis.CAST.Fortran.visitors.cast_visitor import CASTVisitor
 
-from skema.program_analysis.CAST2FN.model.cast import (
+from skema.program_analysis.CAST.Fortran.model.cast import (
     AstNode,
     Assignment,
     Attribute,
