+++ skema/program_analysis/cast_to_agraph.py
@@ -1,9 +1,9 @@
 import sys
 
-from skema.program_analysis.CAST2FN.visitors.cast_to_agraph_visitor import (
+from skema.program_analysis.CAST.Fortran.visitors.cast_to_agraph_visitor import (
     CASTToAGraphVisitor,
 )
-from skema.program_analysis.CAST2FN.cast import CAST
+from skema.program_analysis.CAST.Fortran.cast import CAST
 
 def main():
     """cast_to_agraph.py
