+++ skema/utils/script_functions.py
@@ -11,35 +11,35 @@ from skema.gromet.fn import (
 
 from skema.utils.fold import dictionary_to_gromet_json, del_nulls
 from skema.program_analysis.PyAST2CAST import py_ast_to_cast
-from skema.program_analysis.CAST2FN import cast
-from skema.program_analysis.CAST2FN.model.cast import SourceRef
-from skema.program_analysis.CAST2FN.cast import CAST
-from skema.program_analysis.CAST2FN.visitors.cast_to_agraph_visitor import (
+from skema.program_analysis.CAST.Fortran import cast
+from skema.program_analysis.CAST.Fortran.model.cast import SourceRef
+from skema.program_analysis.CAST.Fortran.cast import CAST
+from skema.program_analysis.CAST.Fortran.visitors.cast_to_agraph_visitor import (
     CASTToAGraphVisitor,
 )
-from skema.program_analysis.CAST2FN.ann_cast.cast_to_annotated_cast import (
+from skema.program_analysis.CAST.Fortran.ann_cast.cast_to_annotated_cast import (
     CastToAnnotatedCastVisitor,
 )
-from skema.program_analysis.CAST2FN.ann_cast.id_collapse_pass import (
+from skema.program_analysis.CAST.Fortran.ann_cast.id_collapse_pass import (
     IdCollapsePass,
 )
-from skema.program_analysis.CAST2FN.ann_cast.container_scope_pass import (
+from skema.program_analysis.CAST.Fortran.ann_cast.container_scope_pass import (
     ContainerScopePass,
 )
-from skema.program_analysis.CAST2FN.ann_cast.variable_version_pass import (
+from skema.program_analysis.CAST.Fortran.ann_cast.variable_version_pass import (
     VariableVersionPass,
 )
-from skema.program_analysis.CAST2FN.ann_cast.grfn_var_creation_pass import (
+from skema.program_analysis.CAST.Fortran.ann_cast.grfn_var_creation_pass import (
     GrfnVarCreationPass,
 )
-from skema.program_analysis.CAST2FN.ann_cast.grfn_assignment_pass import (
+from skema.program_analysis.CAST.Fortran.ann_cast.grfn_assignment_pass import (
     GrfnAssignmentPass,
 )
-from skema.program_analysis.CAST2FN.ann_cast.lambda_expression_pass import (
+from skema.program_analysis.CAST.Fortran.ann_cast.lambda_expression_pass import (
     LambdaExpressionPass,
 )
-from skema.program_analysis.CAST2FN.ann_cast.to_grfn_pass import ToGrfnPass
-from skema.program_analysis.CAST2FN.ann_cast.to_gromet_pass import (
+from skema.program_analysis.CAST.Fortran.ann_cast.to_grfn_pass import ToGrfnPass
+from skema.program_analysis.CAST.Fortran.ann_cast.to_gromet_pass import (
     ToGrometPass,
 )
